requirements_old.txt
APScheduler==3.10.4
Flask==2.0.2
Flask_Cors==3.0.10
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
itsdangerous==2.0.1
langchain==0.0.292
numpy==1.24.4
openai==0.28.0
pytest==7.4.2
python-dotenv==1.0.0
python_dateutil==2.8.2
SQLAlchemy==1.4.0
tabulate==0.9.0
Werkzeug==2.0.3
